removed uneeded IndexElementPipe as an index return is already an iterable. Fixed the RangeFilter so that it goes from 0 to n-1. Added IdVertexPipe so you can stream in ids to retrieve from the graph. Added respective test cases.

+++ src/main/java/com/tinkerpop/pipes/PipeHelper.java
@@ -16,8 +16,8 @@ public class PipeHelper {
         }
     }
 
-    public static <T> int counter(final Iterator<T> iterator) {
-        int counter = 0;
+    public static <T> long counter(final Iterator<T> iterator) {
+        long counter = 0;
         while (iterator.hasNext()) {
             iterator.next();
             counter++;
